+++ udata/api/__init__.py
@@ -1,5 +1,4 @@
 import inspect
-import itertools
 import logging
 import urllib.parse
 from functools import wraps
@@ -18,7 +17,7 @@ from flask import (
 from flask_restx import Api, Resource
 from flask_storage import UnauthorizedFileType
 
-from udata import cors, entrypoints, tracking
+from udata import entrypoints, tracking
 from udata.app import csrf
 from udata.auth import Permission, PermissionDenied, RoleNeed, current_user, login_user
 from udata.i18n import get_locale
@@ -325,7 +324,7 @@ def init_app(app):
     import udata.harvest.api  # noqa
 
     for module in entrypoints.get_enabled("udata.apis", app).values():
-        api_module = module if inspect.ismodule(module) else import_module(module)
+        module if inspect.ismodule(module) else import_module(module)
 
     # api.init_app(app)
     app.register_blueprint(apiv1_blueprint)