+++ udata/core/badges/models.py
@@ -5,9 +5,9 @@ from mongoengine.signals import post_save
 
 from udata.api_fields import field
 from udata.auth import current_user
+from udata.core.badges.fields import badge_fields
 from udata.mongo import db
 
-from .fields import badge_fields
 from .signals import on_badge_added, on_badge_removed
 
 log = logging.getLogger(__name__)
@@ -42,7 +42,11 @@ class BadgesList(db.EmbeddedDocumentListField):
 
 
 class BadgeMixin(object):
-    badges = BadgesList()
+    badges = field(
+        BadgesList(),
+        readonly=True,
+        inner_field_info={"nested_fields": badge_fields},
+    )
 
     def get_badge(self, kind):
         """Get a badge given its kind if present"""