+++ udata/tests/api/test_auth_api.py
@@ -301,7 +301,7 @@ class APIAuthTest:
 
         assert_status(response, 400)
         assert "error" in response.json
-        assert "redirect_uri" in response.json["error_description"]
+        assert "Redirect URI" in response.json["error_description"]
 
     @pytest.mark.options(OAUTH2_ALLOW_WILDCARD_IN_REDIRECT_URI=True)
     @pytest.mark.oauth(redirect_uris=["https://*.test.org/callback"])
@@ -327,7 +327,7 @@ class APIAuthTest:
 
         assert_status(response, 400)
         assert "error" in response.json
-        assert "redirect_uri" in response.json["error_description"]
+        assert "Redirect URI" in response.json["error_description"]
 
     def test_authorization_grant_token(self, client, oauth):
         client.login()
@@ -360,6 +360,8 @@ class APIAuthTest:
         assert200(response)
         assert response.content_type == "application/json"
         assert "access_token" in response.json
+        tokens = OAuth2Token.objects(access_token=response.json["access_token"])
+        assert len(tokens) == 1  # A token has been created and saved.
 
     def test_s256_code_challenge_success_client_secret_basic(self, client, oauth):
         code_verifier = generate_token(48)
@@ -582,23 +584,36 @@ class APIAuthTest:
         )
 
         assert_status(response, 400)
-        assert response.json["error"] == "invalid_grant"
+        assert response.json["error"] == "unsupported_response_type"
 
     @pytest.mark.oauth(confidential=True)
     def test_refresh_token(self, client, oauth):
         user = UserFactory()
-        token = OAuth2Token.objects.create(
+        token_to_be_refreshed = OAuth2Token.objects.create(
             client=oauth,
             user=user,
             access_token="access-token",
             refresh_token="refresh-token",
         )
+        token_same_user_not_refreshed = OAuth2Token.objects.create(
+            client=oauth,
+            user=user,
+            access_token="same-user-access-token",
+            refresh_token="same-user-refresh-token",
+        )
+        other_token = OAuth2Token.objects.create(
+            client=oauth,
+            user=UserFactory(),
+            access_token="other-access-token",
+            refresh_token="other-refresh-token",
+        )
+        tokens_count = OAuth2Token.objects.count()
 
         response = client.post(
             url_for("oauth.token"),
             {
                 "grant_type": "refresh_token",
-                "refresh_token": token.refresh_token,
+                "refresh_token": token_to_be_refreshed.refresh_token,
             },
             headers=basic_header(oauth),
         )
@@ -606,6 +621,26 @@ class APIAuthTest:
         assert200(response)
         assert response.content_type == "application/json"
         assert "access_token" in response.json
+
+        # Reload from the DB.
+        token_to_be_refreshed.reload()
+        token_same_user_not_refreshed.reload()
+        other_token.reload()
+
+        assert tokens_count == OAuth2Token.objects.count()  # No new token created.
+
+        # The access token has been refreshed.
+        assert token_to_be_refreshed.access_token != "access-token"
+        # The refresh token is also updated.
+        assert token_to_be_refreshed.refresh_token != "refresh-token"
+
+        # No change to the user's other token.
+        assert token_same_user_not_refreshed.access_token == "same-user-access-token"
+        assert token_same_user_not_refreshed.refresh_token == "same-user-refresh-token"
+
+        # No change to other token.
+        assert other_token.access_token == "other-access-token"
+        assert other_token.refresh_token == "other-refresh-token"
 
     @pytest.mark.parametrize("token_type", ["access_token", "refresh_token"])
     def test_revoke_token(self, client, oauth, token_type):